requirements.txt
asgiref==3.9.1
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2.4
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.0
dotenv==0.9.9
idna==3.10
Pillow==11.1.0
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
stripe==12.3.0
typing_extensions==4.14.1
urllib3==2.5.0
